requirements.txt
azure-functions
azure-cosmos

pre-commit
black
ruff
pyupgrade
isort
